% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_average_transition_matrix.R
\name{Calculate_average_transition_matrix}
\alias{Calculate_average_transition_matrix}
\title{Title Average Transition Matrix}
\usage{
Calculate_average_transition_matrix(all_transition_matrices)
}
\arguments{
\item{all_transition_matrices}{matrix that contains transition matrix}
}
\description{
This function calculates the average transition matrix with
}
\examples{
# Calculate all transition matrices
all_transition_matrices =   mapply(function(a, b){Calculate_transition_matrix(reclassified_table, a, b)},years[1:(length(years)-1)], years[2:length(years)], SIMPLIFY = FALSE)
Calculate average transition matrix
average_transition_matrix = Calculate_average_transition_matrix(all_transition_matrices)


}
