% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_transition_matrices.R
\name{Calculate_transition_matrix}
\alias{Calculate_transition_matrix}
\title{Title Transition matrices from each year}
\usage{
Calculate_transition_matrix(reclassified_table, year_from, year_to)
}
\arguments{
\item{reclassified_table}{a csv file from the MODIS data extraction}

\item{year_from}{year start}

\item{year_to}{year end}
}
\value{
matrix_transition
}
\description{
This function calculates the transition matrix between two years (year_from,
 and year_to), sourced from the reclassified_table table.
}
\examples{
# Get list of years
years = reclassified_table$Year \%>\% unique() \%>\% sort()
all_lc_matrices = lapply(years, function(year){Get_matrix_by_year(reclassified_table, year)})
}
