% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_jcr_raster.R
\name{Extract_jcr_rasters}
\alias{Extract_jcr_rasters}
\title{Title}
\usage{
Extract_jcr_rasters(jcr_cropped_500)
}
\arguments{
\item{jcr_cropped_500}{the JCR data after resample to 500m resolution}
}
\value{
the table containing X, Y point and water extent of the X,Y point
}
\description{
Title
}
\examples{
all_jcr_table = Extract_jcr_rasters(jcr_cropped_500)
}
