% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_reclassify_rasters.R
\name{Extract_reclassify_rasters}
\alias{Extract_reclassify_rasters}
\title{Title Extract raster classification}
\usage{
Extract_reclassify_rasters(
  modis_cropped_500,
  reclass_table_path,
  reclass_order_path
)
}
\arguments{
\item{modis_cropped_500}{Modis file after cropped with boundary shapefile and reproject to resolution 500m}

\item{reclass_table_path}{path Where the classsification table are stored in .csv}

\item{reclass_order_path}{path Where order of the matrix are stored in .csv}
}
\value{
reclassified_table
}
\description{
The results of this function is a table consist of X,Y point of the MODIS data, classification of the point.
}
\examples{
(require)library(terra)
(require)library(tidyverse)
(require)library(MASS)
(require)library(ggalluvial)
(require)library(rgdal)
(require)library(dplyr)
reclassified_table =
Extract_reclassify_rasters(modis_cropped_500,reclass_table_path,reclass_order_path) \%>\%
drop_na()

}
