% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_LC_bar.R
\name{LCC_bar_plot}
\alias{LCC_bar_plot}
\title{Title LC barplot}
\usage{
LCC_bar_plot(reclassified_table)
}
\arguments{
\item{reclassified_table}{a csv file from the MODIS data extraction}
}
\value{
bar plot
}
\description{
Title LC barplot
}
\examples{
bar_plot = LCC_bar_plot(reclassified_table)
}
