% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_LC_pixel_over_time.R
\name{Pixel_over_time}
\alias{Pixel_over_time}
\title{Title}
\usage{
Pixel_over_time(reclassified_table)
}
\arguments{
\item{reclassified_table}{a csv file from the MODIS data extraction}
}
\value{
an image of land cover over time
}
\description{
Title
}
\examples{
# Countable every year
countTable = reclassified_table \%>\%
 group_by(Year, LC_class) \%>\%
 summarise(Count = n()) \%>\%
 print()
write.csv(countTable, "../results/countable_every year.csv", row.names = FALSE)

#plots the count
ggplot(countTable) +
 geom_area(aes(Year, Count, group = LC_class, fill = LC_class)) +
 scale_fill_manual(values = c("#009e73", "#f0e442", "#d55e00")) +
 theme(legend.position="bottom")

Pixel_over_time(reclassified_table)
}
