% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC Performance.R
\name{project_LC_change}
\alias{project_LC_change}
\title{Title}
\usage{
project_LC_change(
  initial_lc,
  transition_matrix,
  number_of_years,
  reference_final_lc
)
}
\arguments{
\item{initial_lc}{matrix that became the initial year}

\item{transition_matrix}{average transition matrix of n year}

\item{number_of_years}{the k number of time steps}

\item{reference_final_lc}{matrix at t+k}
}
\value{
a number
}
\description{
This function is for check the performance how well the prediction
}
\examples{
proj_2001_2020 = project_LC_change(all_lc_matrices[[1]], average_transition_matrix, 19, all_lc_matrices[[20]])
proj_2001_2020$performance
}
