% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_jcr_raster.R
\name{Load_jcr_raster}
\alias{Load_jcr_raster}
\title{Title}
\usage{
Load_jcr_raster(jcr_database_path, bounding_shapefile_path)
}
\arguments{
\item{jcr_database_path}{= write path where the jcr data store}

\item{bounding_shapefile_path}{= write path where the jcr data store}
}
\value{
jcr_cropped_500
}
\description{
Title
}
\examples{
jcr_cropped_500 = Load_jcr_raster(jcr_database_path,bounding_shapefile_path)

}
