% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_reproject_crop_MODIS.R
\name{Load_reproject_crop_MODIS}
\alias{Load_reproject_crop_MODIS}
\title{Load_reproject_crop_MODIS}
\usage{
Load_reproject_crop_MODIS(
  modis_database_path,
  dataset_selected = "LC_Type1",
  bounding_shapefile_path
)
}
\arguments{
\item{bounding_shapefile_path}{}
}
\value{
modis cropped raster
}
\description{
Load_reproject_crop_MODIS
}
\examples{
modis_cropped_rasters = Load_reproject_crop_MODIS(modis_database_path, dataset_selected, bounding_shapefile_path)

}
